+++ Changes.md
@@ -1,3 +1,4 @@
+* #449 - display playoff times in the finalist schedule
 * #448 - modify finalist schedule times when the start time is modified
 * #427 - run the scheduler and table optimizer in a background thread